+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1 Under development
 
 - Bug #19: Improve language in comments for clarity and consistency in `LanguageChangedEvent.php` and `UrlLanguageManager.php` (@terabytesoftw)
+- Bug #20: Clean up code in tests for better readability and maintainability (@terabytesoftw)
 
 ## 0.1.0 June 20, 2025